Move fhir models into /fhir directory
