+++ src/components/content/medication-history/medication-history.stories.tsx
@@ -9,8 +9,8 @@ import medicationStatement from "@/components/content/medication-history/mocks/m
 import patient from "@/components/content/medication-history/mocks/patient";
 import { CTWProvider } from "@/components/core/ctw-provider";
 import { PatientProvider } from "@/components/core/patient-provider";
+import { MedicationStatementModel } from "@/fhir/models";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
-import { MedicationStatementModel } from "@/models";
 import type { Meta, StoryObj } from "@storybook/react";
 import { rest } from "msw";
