+++ src/components/content/medications-table-base.tsx
@@ -1,8 +1,8 @@
 import { DropdownMenu, MenuItems } from "@/components/core/dropdown-menu";
 import type { MinRecordItem, TableColumn } from "@/components/core/table/table";
 import { Table, TableBaseProps } from "@/components/core/table/table";
+import type { MedicationStatementModel } from "@/fhir/models/medication-statement";
 import { useBreakpoints } from "@/hooks/use-breakpoints";
-import type { MedicationStatementModel } from "@/models/medication-statement";
 import { DotsHorizontalIcon } from "@heroicons/react/outline";
 import { compact, isFunction } from "lodash/fp";
 import { ReactNode, useRef } from "react";
