+++ src/components/content/medications/add-new-med-drawer.tsx
@@ -5,7 +5,7 @@ import {
   medicationStatementSchema,
 } from "@/components/content/forms/medications";
 import { usePatient } from "@/components/core/patient-provider";
-import { MedicationStatementModel } from "@/models/medication-statement";
+import { MedicationStatementModel } from "@/fhir/models/medication-statement";
 import { format } from "date-fns";
 import { ReactElement } from "react";
