+++ src/components/content/medications/other-provider-meds-table.tsx
@@ -1,7 +1,7 @@
 import { MedicationDrawer } from "@/components/content/medication-drawer";
 import { MedicationsTableBase } from "@/components/content/medications-table-base";
+import { MedicationStatementModel } from "@/fhir/models/medication-statement";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
-import { MedicationStatementModel } from "@/models/medication-statement";
 import { sort } from "@/utils/sort";
 import { get } from "lodash/fp";
 import { useEffect, useState } from "react";
