+++ src/fhir/models/condition.ts
@@ -8,8 +8,8 @@ import { CONDITION_CODE_PREFERENCE_ORDER } from "@/fhir/conditions";
 import { findReference } from "@/fhir/resource-helper";
 import { ResourceMap } from "@/fhir/types";
 import { compact, intersectionWith, uniqWith } from "lodash";
-import { formatDateISOToLocal, formatStringToDate } from "../fhir/formatters";
-import { SYSTEM_CCS, SYSTEM_ICD10, SYSTEM_SNOMED } from "../fhir/system-urls";
+import { formatDateISOToLocal, formatStringToDate } from "../formatters";
+import { SYSTEM_CCS, SYSTEM_ICD10, SYSTEM_SNOMED } from "../system-urls";
 import { PatientModel } from "./patients";
 
 export class ConditionModel {
