--- src/models/index.ts
@@ -1,8 +1,0 @@
-export { ConditionModel } from "@/models/condition";
-export { MedicationModel } from "@/models/medication";
-export { MedicationDispenseModel } from "@/models/medication-dispense";
-export { MedicationStatementModel } from "@/models/medication-statement";
-export { OperationOutcomeModel } from "@/models/operation-outcome";
-export { OrganizationModel } from "@/models/organization";
-export { PatientModel } from "@/models/patients";
-export { PractitionerModel } from "@/models/practitioner";
